# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "mygram is a free photo sharing app written in Go. People can share, view, and comment photos by everyone. Anyone can create an account by registering an email address and creating a username.",
        "title": "mygram backend",
        "contact": {
            "name": "gusrylmubarok",
            "email": "gusrylmubarok@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/socialmedias": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all social media with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedias"
                ],
                "summary": "Fetch all social media",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseDataFetchedSocialMedia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create and store a social media with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedias"
                ],
                "summary": "Add a social media",
                "parameters": [
                    {
                        "description": "Add Social Media",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddSocialMedia"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseDataAddedSocialMedia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/socialmedias/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a social media by id with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedias"
                ],
                "summary": "Update a social media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SocialMedia ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Social Media",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateSocialMedia"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseDataUpdatedSocialMedia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a social media by id with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedias"
                ],
                "summary": "Delete a social media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SocialMedia ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessageDeletedSocialMedia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helpers.ResponseMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.AddSocialMedia": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.example.com/johndoe"
                }
            }
        },
...
        "model.AddedSocialMedia": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "the created at generated here"
                },
                "id": {
                    "type": "string",
                    "example": "the social media id generated here"
                },
                "name": {
                    "type": "string",
                    "example": "Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.example.com/johndoe"
                },
                "user_id": {
                    "type": "string",
                    "example": "here is the generated user id"
                }
            }
        },
...
        "model.ResponseDataAddedSocialMedia": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.AddedSocialMedia"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.ResponseDataFetchedSocialMedia": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.SocialMedias"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.ResponseDataUpdatedSocialMedia": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.UpdatedSocialMedia"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.ResponseMessageDeletedSocialMedia": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "your social media has been successfully deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.SocialMedia": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "here is the generated created at"
                },
                "id": {
                    "type": "string",
                    "example": "here is the generated social media id"
                },
                "name": {
                    "type": "string",
                    "example": "Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.example.com/johndoe"
                },
                "updated_at": {
                    "type": "string",
                    "example": "here is the generated updated at"
                },
                "user": {
                    "$ref": "#/definitions/src_modules_socialmedia_model.User"
                },
                "user_id": {
                    "type": "string",
                    "example": "here is the generated user id"
                }
            }
        },
        "model.SocialMedias": {
            "type": "object",
            "properties": {
                "social_medias": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SocialMedia"
                    }
                }
            }
        },
...
        "model.UpdateSocialMedia": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "New Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.newexample.com/johndoe"
                }
            }
        },
...
        "model.UpdatedSocialMedia": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "here is the generated social media id"
                },
                "name": {
                    "type": "string",
                    "example": "New Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.newexample.com/johndoe"
                },
                "updated_at": {
                    "type": "string",
                    "example": "the updated at generated here"
                },
                "user_id": {
                    "type": "string",
                    "example": "here is the generated user id"
                }
            }
        },
...
        "src_modules_socialmedia_model.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "here is the generated user id"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}